+++ CHANGELOG.md
@@ -3,8 +3,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
+### Added
+ - Noob role for users under level 15
+ - Purge command
+ - Superuser shutdown and restart commands
 ### Changed
  - Readme's TODO updated
+ - Balance command to bal
+ - Bal command to b
 
 ## [1.1] - 2021-01-07
 ### Added